# Documentation/reference/openapi.yaml
openapi: 3.0.3
info:
  title: Phantasmon Backend API
  description: >
    API REST du backend Phantasmon, source de vérité des Ghost Pokémon, des échanges et des
    sessions de combat. Consommée exclusivement par le mod client Fabric (aucun serveur
    Minecraft n'appelle cette API).


    Ce contrat décrit l'implémentation réelle (vérifiée le 2026-10-06). Le protocole
    WebSocket (`/ws`) n'est pas décrit ici : voir `websocket-protocol.md`. Exemples et
    comportements détaillés : `rest-api.md`. Catalogue des erreurs : `error-codes.md`.


    Conventions : JSON en snake_case ; erreurs métier structurées (`ErrorResponse`) ;
    les erreurs de format (400) aussi : ERROR_VALIDATION_FAILED, ERROR_MALFORMED_REQUEST.


    Administration (TODO-25) : un admin est un joueur dont le pseudo de dernière connexion figure
    dans le fichier `admins.txt` du backend. Les routes `/admin/*` (hors `/admin/me`) lui sont
    réservées (403 ERROR_ADMIN_REQUIRED) ; il peut aussi lire et modifier le PC de n'importe quel
    joueur avec les routes Pokémon habituelles.
  version: 0.1.0
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Développement local
...
security:
  - bearerAuth: []
paths:
...
  /hub/anchors:
    get:
      tags: [Hub]
      summary: Hub Anchors d'un serveur et d'une dimension, partagés par ses joueurs (D-30)
      parameters:
        - name: server_fingerprint
          in: query
          required: true
          schema: { type: string, maxLength: 128 }
        - name: dimension
          in: query
          required: true
          schema: { type: string, maxLength: 128, example: "minecraft:overworld" }
      responses:
        "200":
          description: Anchors triés par nom
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HubAnchor"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Hub]
      summary: Crée un Anchor de l'appelant (un par hub, idempotent via request_uuid)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [request_uuid, hub, name, server_fingerprint, dimension, origin, yaw]
              properties:
                request_uuid: { type: string, format: uuid }
                hub: { type: string, maxLength: 32, example: global, description: Nom du hub où mène l'Anchor (D-35) }
                name:
                  type: string
                  minLength: 3
                  maxLength: 32
                  description: Lettres, chiffres, espaces, - ou _, sans espace au début ni à la fin ; unique par serveur (casse ignorée)
                server_fingerprint: { type: string, maxLength: 128 }
                dimension: { type: string, maxLength: 128 }
                origin:
                  type: object
                  required: [x, y, z]
                  description: Position des pieds du créateur
                  properties:
                    x: { type: number, minimum: -30000000, maximum: 30000000 }
                    y: { type: number, minimum: -2048, maximum: 2048 }
                    z: { type: number, minimum: -30000000, maximum: 30000000 }
                yaw: { type: number, description: "Orientation Minecraft quelconque, arrondie au quart de tour" }
      responses:
        "201":
          description: Anchor créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubAnchor"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: ERROR_HUB_ANCHOR_QUOTA, ERROR_HUB_ANCHOR_NAME_TAKEN ou ERROR_HUB_ANCHOR_OVERLAP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /hub/anchors/{uuid}:
    delete:
      tags: [Hub]
      summary: Supprime un Anchor (créateur ou admin)
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Supprimé
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /hubs/{name}/schematic/file:
    get:
      tags: [Hub]
      summary: La construction d'un hub (.schem ou .litematic, NBT gzip)
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Le fichier tel quel ; ETag = SHA-256
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/Error"
  /admin/hubs:
    post:
      tags: [Admin]
      summary: Crée un hub et son dossier de schematics (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, length, width, height]
              properties:
                name: { type: string, pattern: "^[a-z0-9_]{2,32}$", description: Mis en minuscules }
                length: { type: integer, minimum: 3, maximum: 64, description: Vers l'avant de l'Anchor (z) }
                width: { type: integer, minimum: 3, maximum: 64, description: En travers (x) }
                height: { type: integer, minimum: 3, maximum: 64 }
      responses:
        "201":
          description: Hub créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hub"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /admin/hubs/{name}:
    delete:
      tags: [Admin]
      summary: Supprime un hub et ses Anchors ; archive son dossier (admin)
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  hub: { type: string }
                  anchors_deleted: { type: integer }
                  archived_as: { type: string, nullable: true }
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/hubs/{name}/reload:
    post:
      tags: [Admin]
      summary: Relit le dossier de schematics d'un hub (admin)
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Hub relu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hub"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Erreur métier structurée (voir error-codes.md)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error_code, details]
      properties:
        error_code:
          type: string
          example: ERROR_LEGALITY_EV_TOTAL_EXCEEDED
        details:
          type: object
          example: { total: 528, max: 510 }
...
    HubAnchor:
      type: object
      required: [uuid, owner_uuid, hub, name, server_fingerprint, dimension, origin, yaw, size, created_at]
      properties:
        uuid: { type: string, format: uuid }
        owner_uuid: { type: string, format: uuid }
        hub: { type: string, example: global }
        name: { type: string, example: "Place du marché" }
        server_fingerprint: { type: string }
        dimension: { type: string, example: "minecraft:overworld" }
        origin:
          type: object
          required: [x, y, z]
          properties:
            x: { type: number }
            y: { type: number }
            z: { type: number }
        yaw: { type: integer, enum: [0, 90, 180, 270] }
        size:
          $ref: "#/components/schemas/HubSize"
        created_at: { type: string, format: date-time }
    HubSize:
      type: object
      description: Taille du hub en blocs — x largeur, y hauteur, z longueur (vers l'avant de l'Anchor)
      properties:
        x: { type: integer, example: 21 }
        y: { type: integer, example: 21 }
        z: { type: integer, example: 21 }
    Hub:
      type: object
      required: [name, size]
      properties:
        name: { type: string, example: global }
        size:
          $ref: "#/components/schemas/HubSize"
        schematic:
          type: object
          nullable: true
          properties:
            name: { type: string }
            format: { type: string, enum: [SCHEM, LITEMATIC] }
            sha256: { type: string }
            bytes: { type: integer }
...